README: Replace stable-2.15 usage with master's
As this is the master branch. Before this change, the version of that
README.md must have been based on the stable-2.15 one. This was most
likely because stable-2.16 did not exist back then, thus using
stable-2.15 values to initialize a master version of it.

Use the currently latest bazlets commit in master to initialize this
README (in/for master) accordingly.